Adecuacion de tipo de dato para el objeto de conductor y perosnaVO se tenia mapeado el id como integer, ae genera dao para consultar transportistas sin vehiculo, y adecuacin para el alta de os

+++ siye-api/src/main/java/mx/com/teclo/siye/negocio/service/catalogo/TblCatalogosService.java
@@ -19,5 +19,5 @@ public interface TblCatalogosService {
 
 	public Boolean actualizaPersona(PersonaVO personaVO) throws NotFoundException, BusinessException;
 
-	public PersonaVO findPersona(Integer idPersona);
+	public PersonaVO findPersona(Long idPersona);
 }
